DeveloperGuide, UserGuide: use full xrefstyle
By default, the cross-reference text when cross-referencing another
section, figure, table etc. will be the title of that section, figure,
table etc.

For example, the following cross-reference in the DeveloperGuide:

    <<architecture>>

will be replaced with:

    Architecture

which is the name of that section.

However, in a printed or non-interactive medium, cross-reference
hyperlinks do not work. As such, it would be more useful to provide the
section/figure/table number as well, so that readers will be able to
quickly locate the reference target by its number, rather than needing
to manually find its name in the document.

As such, lets set the xrefstyle of the UserGuide and DeveloperGuide to
`full`. With this change, the <<architecture>> cross-reference would be
replaced with:

    Section 2.1, "Architecture"

As the xrefstyle feature[1] was only introduced in asciidoctor 1.5.6[2],
we'll need to upgrade our asciidoctor plugin as well.

[1] http://asciidoctor.org/docs/user-manual/#customizing-the-cross-reference-text
[2] https://github.com/asciidoctor/asciidoctor/commit/8e80317dee56483e48a07451bff0cd14f0bfd59c