AddressBook#resetData(ReadOnlyAddressBook): Reorder logic
resetData(...) performs the following operations in order:
* Setting new values for UniquePersonList `persons` and UniqueTagList
`tags`
* Synchronizing the tags in `persons` with `tags`

This order of execution requires Person to be mutable so that the tags
in `persons` can be synchronized with `tags`.

As part of making Person to be immutable, let's reorder the logic by
performing the synchronization before setting new values for
UniquePersonList `persons`.

Also, let's update syncMasterTagListWith(Person) to take in a
ReadOnlyPerson instead since we are making Person immutable and should
no longer set a Person's tags. As a result,
syncMasterTagListWith(Person) will have to return a Person object.
Let's also update the method name and header comments to reflect these
changes.