Developer Guide: Remove horizonal lines in undo/redo explanation
The explanation of the undo/redo feature in the DG contains a number of
horizontal lines which are used to show visual hierarchy of the content.
This part of DG is included in the sample Project Portfolio Page (PPP)
and that results in the said horizontal lines interfering with the
visual hierarchy of the sample PPP.

Let's remove the offending horizontal lines and use headings and bullet
points to show the visual hierarchy instead.