[#775] Developer Guide: specify target user profile and value proposition (#776)
One needs to have a clearly defined target user profile and a value
proposition for the app before other requirements are defined. Both
are not specified in the Developer Guide.

Let's add a sample target user profile and a value proposition to the
DG as a separate appendix named 'Product Scope'.