Remove ReadOnlyPerson
ReadOnlyPerson provides a read-only immutable interface for Person.

Person is already immutable. This removes the need for ReadOnlyPerson.

Let's replace all instances of ReadOnlyPerson with Person.

Mechanical changes:
1. Delete ReadOnlyPerson. Ensure that all
"import seedu.address.model.person.ReadOnlyPerson;" statements are
deleted.
2. Use the replace all functionality to replace
"? extends ReadOnlyPerson" with "Person" for all text files.
3. Use the replace all functionality to replace
"ReadOnlyPerson" with "Person" for all text files.

At this point, the code base will not compile as the
"import seedu.address.model.person.Person;" statement is missing in
some files. You will have to manually add this line to those files,
using your IDE's tooling if applicable. On Eclipse, you can press
Ctrl + Shift + O to do so.

Furthermore, in AddressBook#updatePerson(Person, Person), the code will
not compile because a variable is re-defined. This will also need to be
manually fixed as well.

Finally, rewrap some lines which have become too short as a result of
the find-and-replace in steps 2 and 3.

This commit was performed using IntelliJ.