Person: Remove copy constructor
This copy constructor existed for the purpose of defensive copying.

Person is immutable. There is no need for defensive copying.

Let’s remove it.

Some methods depended on Person's copy constructor to throw
NullPointerException if the given Person object is null. Let's teach
those methods to verify that the given Person object is non-null.