UndoableCommand: add preprocessUndoableCommand() method
UndoableCommand does not have a method to preprocess and generate
the values required for the execution of the command.

This leads to the commands handling the preprocessing step within the
UndoableCommand#executeUndoableCommand() method. This violates SRP,
as each method is only supposed to do only one thing.

Let's add a new method preprocessUndoableCommand() for these commands
to run their preprocessing step on.