docs: add DummySearchPage.html
BrowserPanel loads the Google search page whenever a person card is
selected. The tests written for this feature verifies that the loaded
web page is indeed the Google search page with the person's name.

When many forks of this project are running automated tests, too many
search requests are sent to Google, causing us to exceed the search
service’s rate limits. This is an abuse of Google’s search service.
Also, this causes our tests to fail as the loaded web page will be
redirected to a captcha page.

As part of updating BrowserPanel to no longer load from Google search
page, let’s create a DummySearchPage.html in our docs folder. This file
will be automatically deployed onto GitHub Pages whenever we push a
commit to the master branch. GitHub Pages serve static web pages which
requires much less processing power, thus it is more unlikely for us to
exceed the search service’s rate limits and also more unlikely for us
to be redirected to a captcha page.