XmlAdaptedPerson: update equals(Object) method
XmlAdaptedPerson#equals(Object) does an equality check by calling the
equals(Object) method for each of its field.

As it is possible for some of these fields to be null due to a
corrupted .xml save file, directly calling equals(Object) for any
of these fields is not safe as it may lead to NullPointerException.

Let's update XmlAdaptedPerson#equals(Object) to call the safer
Objects.equals(Object, Object) method for the necessary fields.