AddressBook: remove master tag list
We keep track of a master tag list which holds all the tags in
AddressBook.

The tag list is not implemented correctly. The original intention was
for tags in Person to point to the tags in the master tag list.
However, implementing it in this way may be too complex for new
developers to grasp, as discussed in #753. Furthermore, from the
discussion in #794, the tag list in its current state is not needed.

Let's remove the master tag list in AddressBook.