Person: add equality method for identity fields
Person class contains only one equality method Person#equals(Object) 
which checks if all fields of the Person except tags are identical.

Two Persons of the same name that have at least one other equivalent 
identity field (phone or email) but different data fields (address and/
or tags), is considered different even though they are actually the same 
person.

Let's add another equality method, Person#isSamePerson(Person), for 
identity fields in Person class. Adding an equality method for identity 
fields will allow detection of such cases which may be unintuitive to 
users.

Following this change, some tests throw DuplicatePersonException when 
they try to add the default person using PersonBuilder to a model that 
already contains TypicalPersons#ALICE. This is because the default 
person in PersonBuilder has the same values for the name and phone 
fields as ALICE. To prevent that, let's change the value of the phone 
field for ALICE.

Furthermore, PersonListPanelTest#createXmlFileWithPersons(int) throws a
DuplicatePersonException as all Persons created have the same name and 
email. To prevent that, let's change the format of the name and phone 
creation such that each Person is unique.
r