SelectCommandTest: verify model state as well
SelectCommand, when executed, should:

 (1) Raise a JumpToListRequestEvent (if the provided selection index is
     valid), and

 (2) Not modify the model at all.

SelectCommandTest verifies (1), but does not verify (2).

Given that in the actual application, LogicManager will pass the Model
to SelectCommand, we should verify that SelectCommand behaves as
expected and does not modify the model.

As such, let's make SelectCommandTest verify the model state as well.
This is done by rewriting it to use
`CommandTestUtil.assertCommandSuccess(...)` and
`CommandTestUtil.assertCommandFailure(...)`, for consistency with the
other command tests.