XmlAdaptedPerson#toModelType(): remove shadowing of fields
The method body references the fields of XmlAdaptedPerson (`name`,
`phone`, etc.), but also shadows them later on by declaring local
variables with the same name.

Such mixing of instance fields and local variables within the same
method is out of the ordinary in the code base, and can be accidentally
mis-read if the reader is not paying attention.

Let's take the path of least surprise by renaming the local variables
such that they do not shadow XmlAdaptedPerson's fields.