ModelManager: remove synchronization from methods
Synchronized methods are only useful when multiple threads could access
the same object at the same time[1].

However, when our application is run, only the JavaFX application thread
should have access to the model. When running GUI tests, the main thread
(running the tests) and the JavaFX application thread will both have
access to the model, but the synchronization implemented internally in
TestFX should ensure that only one thread accesses the model at any
given time.

As such, the model does not require synchronization. If any of the above
conditions do not hold, it is a bug and should be fixed directly, not
papered over by adding locks. Depending on the JVM's implementation
details and whether is having a good or bad day, such careless use of
synchronization could lead to a performance penalty[2].

So, let's remove the synchronization from ModelManager's methods, and
remove the synchronization requirement from its javadoc.

[1]: But even then, they're not a silver bullet.
[2]: https://stackoverflow.com/a/8524882