ResultDisplay: migrate away from EventsCenter
When MainWindow gets a CommandResult or exception from calling
Logic#execute(String), it would notify ResultDisplay of the result by
posting a NewResultAvailableEvent to EventsCenter.

As we are going to remove EventsCenter in the future, let's migrate
ResultDisplay away from EventCenter.

Rather than having this event-posting-and-receiving arrangement, let's
add a ResultDisplay#setFeedbackToUser() method to ResultDisplay, and
teach MainWindow to call this method when it receives a CommandResult
from Logic#execute(String).