ExitCommand: use CommandResult to signal MainWindow to exit
To make the application exit, ExitCommand will post a
ExitAppRequestEvent to EventsCenter. MainWindow will then respond by
hiding itself (and its HelpWindow if open). Since JavaFX will exit when
all windows are closed, this will cause the application to exit.

As we are going to remove EventsCenter in the future, let's migrate this
mechanism away from EventsCenter.

Let's add an `exit` flag to CommandResult, and teach ExitCommand to use
this flag to signal to MainWindow to exit. Teach MainWindow to respond
to this flag by hiding itself (and its HelpWindow if open).