Address book storage: migrate away from EventsCenter
When the model's address book is modified, ModelManager will post an
AddressBookChangedEvent to EventsCenter. StorageManager will then
respond by calling StorageManager#saveAddressBook(...) to save the
address book.

If an error occurs while saving the address book, StorageManager will
post a DataSavingExceptionEvent to EventsCenter. UiManager will then
respond by showing an alert dialog.

As we are going to remove EventsCenter in the future, let's migrate
these mechanisms away from EventsCenter.

Let's teach LogicManager to listen for invalidation events from
ModelManager's AddressBook. If an invalidation event occurs (which
indicates that the AddressBook was likely modified), LogicManager will
then call Storage#saveAddressBook(...) after command execution. If an
error occurs, this will then be sent to the caller via exceptions. The
end result is that the control flow is much clearer and explicit in the
code.

Note that the new implementation differs in the handling of address book
storage errors: previously an error dialog would be shown. Now, the
error would be reported by the ResultDisplay. This approach is chosen as
its much simpler -- we can simply reuse the same battle-tested code
paths for handling command errors. Furthermore, the overall goal is
still achieved as the error is still reported to the user.

The Developer Guide describes in detail how address book saving works
under the section "Events-Driven nature of the design". As such, we'll
need to update it. Since at this point most of the codebase is not using
EventsCenter, let's rewrite the entire section into one that describes,
in high-level detail, how architectural components interact with each
other.