LO-OCP: update to match the current codebase
The LO still refers to the names of the old classes (`Logic`, which was
renamed to `LogicManager`, and `Parser`, which was renamed to
`AddressBookParser`). Update them to the new names.

Also, this LO refers to the old LO-DIP, which had an exercise on
applying DIP to the `Logic` class by modifying its design.  However, in
the current codebase, DIP has already been applied to the design of the
`LogicManager` class, and LO-DIP was updated in 5f012189 (LO-DIP: update
to match the current codebase, 2019-06-11) to account for that by
replacing the exercise with a statement on how the current design is
already an application of DIP. Let's update LO-OCP in a similar manner.