template header: replace AB-3's nav item with AB-3.5
As AB-3.5 was derived from the AB-4 codebase, it is currently occupying
AB-4's position in the site navigation bar.

However, AB-3.5 is not meant to replace AB-4, but to replace AB-3 in the
future.

As such, let's modify the site navigation bar such that AB-3.5 replaces
AB-3, and restore AB-4's navigation item. This sets up the site
navigation bar nicely for AB-3.5 to take over AB-3 in the future.