Replace all references to AB3.5 with AB3
The AB-3.5 repo is now ready to become the new AB-3. Compared to the old
AB-3, the AB-3.5 codebase is derived from the AB-4 codebase, and thus
shares all of its improvements, and allows us to more easily port over
changes from AB-4.

To allow the replacement to happen, let's replace all references to
AB-3.5 with AB-3.

This is done by grepping the codebase for:

 * `Level 3\.5`
 * `level35`
 * `AB3\.5`
 * `AB-3\.5`
 * `\b3\.5\b`

After this conversion, the old `se-edu/addressbook-level3` repo on
GitHub should be renamed to something else, while the
`se-edu/addressbook-level35` repo on GitHub (this repo) should be
renamed to `se-edu/addressbook-level3`.