Migrate to Jekyll, GitHub Actions
This project uses AsciiDoc for doucmentation. The courses that use
this project no longer needs section numbering and content inclusion
features in documentation. That opens the door for going back to
Markdown, which is more widely used and simpler.

This project also uses Travis/AppVeyor for CI.
With the recent release of GitHub Actions, there is no need to depend
on external CI services.

Let's do the following changes:
* Migrate documentation from AsciiDoc to Markdown, and use Jekyll to
  generate the project website. The migration was done mostly by
  @j-lum, with minor cosmetic/content tweaks added by @damithc
* Migrate to GitHub Actions (from Travis and AppVeyor).

As a side-effect, we have to drop support for Coveralls,
because the current Coveralls set up depends on Travis.
We expect to bring back Coverall support in future.

We decided to do both major changes at the same time in order to
avoid some overlaps, and also snuck in some documentation tweaks
because almost every documentation line is touched by the
asciidoc to Markdown conversion anyway.