Upgrade to Gradle 6.9.2
While newer versions of Gradle are available, AB3 still uses Gradle
5.2.1. The archiveName property is also deprecated and replaced
with the archiveFileName property while the shadow plugin has
also been updated.

Let's upgrade Gradle to 6.9.2 as an intermediate version between 5.2.1
and 7.4.2, resolve the deprecation and update the shadow plugin to the
latest version that is compatible with Gradle 6.9.2.