Add a null check to LogCenter.getLogger(Class<T>)
LogCenter.getLogger(Class<T> clazz) creates a Logger named
after the given class. If the clazz is null, it is supposed to
create a Logger with an empty String as the name.

The current implementation calls Logger.getLogger("")
instead of LogCenter.getLogger(""), which results in the Logger
object not having the handlers that are added inside the
LogCenter.getLogger(String) method. This means log messages
written using such a logger will not appear in the log files.

However, the current codebase has no use case for passing a
null object to this method. As per the YAGNI principle,
we should not implement that ability until we actually need it.

Hence, instead of correcting the said error, let's add an assertion
to ensure null is never passed to this method, eliminating the need
for the buggy code path in the first place.