Update site configuration for project fork
The configuration file in project still set to upstream values,
referencing the original project name and URLs.

These configuration need to reflect the forked project to ensure the
documentation and site setup align with the project.

Update the files to change the product name.

The docs.yml is essential for the correct routing of links in deployed
site while site.json affect the local site.

These updates ensure the documentation and deployed site reflect the
correct project name.