Update DeveloperGuide.md
The information about the user stories and use cases are not up to
date with the planned changes.

This makes the planning phase much more difficult as the information
is not consolidated into a single source.

Let's update the developer guide to consolidate all the plans into
one place to make it easier for developers to follow the guide while
the app is being developed.

This makes the plan more concrete and easy to follow, with less room
for ambiguity.