Change command input to be lowercase
The original address book does not understand the command when the user types "List" or "DELETE" or "LiSt", etc

The new change converts the input of the command (ie. ONLY the command word eg. "add", "list" and not the rest of the input) to lowercase so that TP can undestand the input even if user types in uppercase