Add ListCommandParser class
In AddressBookParser class, the list command does not check if there are any additonal arguments after the command "list". As such, the students' details would be successfully listed if user input was something like "list all"

Let's add a ListCommmandParser class to check if the user input is valid for the list command (ie. just the word "list"). Else, TP will send a message to user telling the user of the correct format