Add user profile and stories to Dev Guide
Added details to the Developer Guide, including the following:

- Defined the target user as independent Geriatricians managing elderly
  patients.
- Added the value proposition highlighting how the app addresses the
  needs of Geriatricians that existing software doesn't fulfill.
- Included user stories for different stages of user experience:
  - First use (new user)
  - Second use (regular user)
  - Tenth use (novice user)
  - Hundredth use (experienced user)
- Marked specific user stories for MVP and final versions.

These updates help clarify the purpose and functionality of the app for
developers.

with '#' will be ignored, and an empty message aborts the commit. # # On
branch feature-update-DG # Your branch is up to date with
'origin/feature-update-DG'. # # Changes to be committed: #
modified:   DeveloperGuide.md #