Add DateUtil and DateOfBirth classes
Currently there is no class to handle dates, hence there should
be a utility class to validate and manage date format.

A DateUtil class is added to help check if given dates follow a
specific yyyy-MM-dd format. A DateOfBirth class is also added,
which utilises the DateUtil class, that will serve as a parameter
for the Person class.