Update Person classes and add new class for parameters
The DateUtil class was updated to contain only methods to be used
and not for storing assigned dates. Gender and Nric classes are
added to contain logic for the supported parameters when adding
patients.

The Person class was updated to store the new parameters and the
logic of checking whether two person are the same was changed to
compare their Nric instead