Create order package files
Extension of AB3 in v1.2 includes the addition of orders to be created and
added under a person in the address book. In this case, the order package
and relevant classes must be created.

Creating the order package and relevant classes allows us to extend AB3 by
associating persons in the address book to their orders. Orders include
details of order item, count and delivery date, represented as respective
classes following AB3 OOP style.

Refactoring other files was necessary due to inclusion of orders and an order
list into Person.java, affecting the Person.java constructor.

2 todos were commented to mark further implementation needed for json
adaptation for storage, as well as in PersonBuilder.java that allows for
Order creation on Person creation. Implementation only allows for orders
to be created after Person is added.

No commands for creating and operating on person's orders have been
implemented, and will be followed up with in further updates. Similarly,
test cases have yet to be implemented.