Create Event Class
Event class has been created, supporting the creation of events with and
without descriptions so that the description field is optional for the
user to fill in.

Each of its parameters have their respective classes. Below are the
parameters and the data type that they store:
* Name, Location, Description -> String
* startTime, endTime -> LocalTime (using HH:MM format)
* Date -> LocalDate (using yyyy-MM-dd format)

Currently, Date and Time classes' constructors may throw a
DateTimeParseException to catch the user's bad inputs that do not follow
the required format.

Currently UniqueEventList::setEvent is implemented but is not in
use until we decide to add the edit function.