Modify Plans after testing
There is a weird behaviour found while testing. The plan IDs are
not initialised within the constructor and all plans have a separate
variable called ID.

When these plans are created using Insurance plan as the object,
the ID defaults to 0. The equals method then works weirdly because
then all plans created through polymorphism has the same ID.

Let's make the ID global variable only appear inside of the
InsurancePlan object with an invalid id and the variable is then
declared differently in each class.

This allows the equals method to work as intended in future with
polymorphism which should allow any checking of duplicate plans to
be more correct