Exit command: display exit message
The feedbackToUser is passed to the resultDisplay UI, but the
application closes too quickly for the feedback to be rendered to the
user.

Let's add a 3 second delay upon exit.

Running the delay (and handleExit behaviour) on a separate concurrent
thread allows the feedbackToUser to be rendered by FXML.

The duration of the delay (3 seconds) is arbitrarily selected, and can
be modified later based on user feedback.

Let's disable the commandBox.

Disabling the commandBox prevents new commands from being executed. As
the handleExit behaviour is running concurrently with the application's
command execution, a command's behaviour might be interrupted before it
completes all its steps. For example, after exiting, the user deletes a
person record, but the application closes before the deletion is saved
by the storage module.