Move helper function arePrefixesPresent() up to ParserUtil.java
Since now the find command demands at least one
name to be specified, arePrefixesPresent() will
be needed by both find and add. We need to avoid
code duplication and make sure it is only defined
once.

Let's move it to ParserUtil.java so that parsers for both add and find can access it.