Create new Policy classes
Policy:
An abstract class for all Policies.

PolicyType:
An enum to represent all Policy types.

LifePolicy, HealthPolicy, EducationPolicy:
The respective Policies for each Policy type.

PolicyMap:
A collection of Policies.
Unlike Tags, Set<Policy> is not used as we would like to override the toString
method to prettify how a set of policies is printed. We would also like to add
custom logic to the add method, as two policies of the same type cannot be added
to the same collection, which is not supported if Set<Policy> was used.