Add ClearCommandParser and update ClearCommand
ClearCommand currently clears the entire address book without
considering any specifications by the user.

This will not support our MVP specifications for the clear command to
take in tags or filters to clear by.

Let's add a ClearCommandParser to provide future parsing support when
we add the functionality for ClearCommand in the future.

The ClearCommandParser uses similar logic to the FindCommandParser, except
throwing an error when the arguments is empty. This is to allow the original
command to still take place.