Fix potential concurrency bug
The commandBox is disabled in the concurrent thread. This allows some
time after an exit command when another command could be runned
illegally.

Let's move the disabling of commandBox out of the concurrent thread into
the main thread.

Disabling the commandBox in the main thread ensures that the commandBox
is disabled before the current command finish processing, in turn
disabled before another command can start processing.