Allow nickname optional
Add command currently makes the -nn tag compulsory, if not mistaken. For
empty nicknames, the GUI shows them as an empty line below the email. To
allow nicknames to be optional when writing the add command, and for the
nickname component to only be shown if the nickname is not an empty
String, let's

* Change the nickname component in fxml file as a FlowPane rather than a
  Label
* Change the code in AddCommandParser.java where
  arMultimap.getValue(PREFIX_NICKNAME).orElse("") rather than .get() so
  that an empty Optional object returned will be seen as a empty String
  return to alow for parsing in parseNickname function

This are the main changes to allow nickname to be an optional
requirement in Add Command.