Update Delete Command
The current Delete command only supports the deletion of a student
via specifying the index of the contact to be deleted in a list.
We want to be able to delete a specific student contact through
specifying their full name.

Let's:
Modify DeleteCommand and DeleteCommandParser to support accepting a
Name to be deleted.
Additionally, we modify Model and ModelManager such that we can
find persons by their names.