Update valid contact
Restriction on contact states that it should contain only numbers and should be at least 3 digits long.

Since our target user is a clinic administrative in Singapore, all phone numbers are expected to be 8 digits long.

Let's update the condition such that all phone numbers are 8 digits long and update all the test cases and test persons as well.