Add new field under a person, income
There is no current field to represent the income of a saved contact.
This information is important to know when serving customers.

Creating this field allows for contacts to have an associated income
level.

As of v1.2, the `add` and `edit` commands can be used to add a new
contact with the income field being mandatory, and edit an existing
contact's income level accordingly.