Adjust tests + change equality logic
Tests broke because of the new definition of person and its constructor.
In addition, changes to equality logic broke some existing tests.

Adding income as a parameter fixes the tests.
Adjusting the equality logic tests, such that a person is equal to
another person if their
- name
- email
- address
are equal.

We will need to discuss criteria for two persons being equal to each
other and adjusting our test logic accordingly.