Add grade level field to Person
The user, who is piano teacher in Singapore, does not have a way to
check the grade level of the students, while doing so would benefit
them by featuring additional data.

Add grade level to Person to allow the user to store the grade level

The commit not only includes the addition of the grade level field to
the Person class, but also the necessary changes to the commands,
parsers, and storage classes to support the new field. The commit also
includes the necessary changes to the UI to display the grade level.

Additionally, the commit modifies User Guide to include the new field
and its usage.

Let's add the grade level field to the Person class to allow the user
to store the grade level of the students.