Create Unit Tests for InsurancePlanFactory.java
There are no test cases to ensure the robustness of the insurance
plan factory.

It becomes harder to catch regressive errors in future development
work.

Let's add robust testing to the class to ensure that it works as
intended in future.

This will help catch regression error faster and more efficiently.