Add Appointment Command
Added Appointment Command. This Command adds an
appointment to a client.

An Appointment command comprises of a Date, From
and To classes. Which represent the respective
fields of an appointment with a client.

Added AppointmentCommandParser to parse the
correct arguments for an appointment command.

Modified relevant tests and other supporting files
such as JsonAdaptedPerson to support the addition of
the new appointment label.