AddressBook: Add ProjectListPanel and ProjectListCard
Add ProjectListCard.fxml and ProjectListPanel.fxml to resources/view.
Add ProjectCard class and ProjectListPanel class to seedu.address.ui.

Fix bug in logic.getFilteredProjectList() which returns an empty ObservableArray by temporarily replacing it with a static list containing a single project.

The ProjectListPanel displays projects by retrieving the data from an ObservableArray. The bug prevented the panel from populating, so a static project is used as a placeholder for now.