Add View Command Functionality
Current iteration of AB3 does not support the accessing of certain
contacts to view all information / information that may not be
displayed on the Person Card component.

As such, the user may be unable to access or even view other
critical information of their contact of interest.

Let's incorporate that functionality, allowing users to view
ALL contact information of the specified contact with the specified
telegram handle.