Modify GUI view to match first mockup
Original GUI was plain dark gray and rigid.

Change background colors to some stackpanes and
change font size and font for some text. Also
added comments to identify what was changed from
original and also to label which components were
responsible for what.

When window is resized to become larger, a gray
box appears under the company-view, unable to
target it currently.