Modify DeleteCommand to delete by name, phone and email
DeleteCommand only deletes using index. This is not convenient for users
as they need to find the index of the contact first before deleting
the contact.

Let's modify DeleteCommand to delete using name, name and phone or name
and email. This commit does this by the following:
* Modifies DeleteCommandParser to return DeleteCommand whose argument
is name, name and phone, name and email, or index.
* Allows DeleteCommand to be initialised by any of the argument
combination as mentioned by using different constructors.
* Uses Optional to handle the null objects more elegantly.
* Implements this flow to delete using name and phone, or name and
email: filter contact list by predicate, get
Person to delete if no duplicates, get Person index in contact list,
and returns CommandResult as previously implemented.
* For name, if there are duplicate names, will ask user to input name
and phone or name and email.
* If there is no contact in the filtered list, it will return a no
contact found exception.