Add FullNameMatchesPredicate and fix checkstyle
There is no predicate that matches full name exactly for FilteredList.

Let's create a FullNameMatchesPredicate class that checks if the full
name of a contact matches exactly with the full name given by user,
so that the user can specifically delete the contact with matching
full name.

Also fix unusable imports.