Update emergencyContactCommand tests
Test cases fail due to new changes in emergencyContactCommand logic in
which emergency contacts now cannot be overridden.

Update add emergency command test case to use different Person with no
emergency contact to begin with. Remove delete emergency contact test
which will be implemented in the future along with the delete function.
Add new test case to test that attempts to override an emergency
contact results in a thrown exception.

To ensure that test cases are up to date and able to ensure the
functionality of the product.