Edit logic in execute() method for EmergencyContactCommand
Current logic checks if emergency contact name field or emergency
contact number field is not empty, and if either is not empty it will
throw the override exception.

Both should be empty before disallowing the user to override the
emergency contact.

Maintains the integrity of emergency contact such that it has to have
both a name and number.