Use different colors for roles
All roles are rendered as blue labels in UI.

Different colors allows users to identify different contacts with
different roles easier.

A private method getRoleLabel is added to set id of the label
corresponding to its role name. Different colors are added in DarkTheme
.css to apply on labels with different ids.

A label needs to be associated with an id in order for CSS selectors to
locate the element. Hence, each role name needs to be mapped to an id.