Implement iew GUI changes
Currently, iew updates the command output segment, while keeping
the list of contacts visible.

As the number of fields for each contact grows, the lack of vertical
space in that segment means the view command serves little purpose in
allowing the user to view all the information of the specified contact.

Let's hide the list of contacts, and display a new page below the command
output section that showcases all the existing (and future) information
of the specified contact whenever the iew command is called.