Fix error message issues
Claims and Insurance uses the same Exception when throwing errors.

This makes it difficult for the developer to throw different
error messages when either one fails as both are dependent on the
same Exception object type.

Let's add separate Exception type objects to both such that
different error messages can be shown to the user in case of wrong
input or execution.

This will also free up namespace cluttering in future development.