feat: Handle person data with illegal value
This commit update the way person data with illegal value is handled.
Originally, one or more illegal value will result in
the whole data file being treated as corrupted.
The app will then start on a clean slate.

Now, if a person data contain(s) illegal value(s),
only that particular person data will be discarded.
Person data with no issue will be kept and read as usual.
However, if any other part of the file is corrupted,
it will still discard the whole file
and start on a clean slate as per previous version.

This update improves the usability of the application
and minimises data loss.