Add remark field to 'add' and 'edit' + minor fixes
The remark command is the only way to add and edit remarks to a person.
Using this separate command from `add` and `edit` is not only unintuitive
but also slows down users.

Adding the functionality of the `remark` command to `add` and `edit`
will allow us to improve on these 2 aspects.

As of v1.3, the `add` and `edit` commands can be used to add a new
remark, using the `rn/` flag. And for `edit` the `ra/` flag can be used
to append a new remark onto existing remarks. These flags remain
optional.

Aside from this the following minor changes are made:
- Add error message for concurrent `rn/` and `ra/` messages being
  specified in `Messages` class.
- Update format for the status bar message that is shown to users when a
  person is added or edited to include remark in `Messages` class.
- Add remark constraints message in `Remark` class.
- Add validity test for remark in `Remark` class.
- Add method to combine multiple remark in `Remark` class.