Add ModuleCommand feature
Currently, only the GradeCommand supports adding a module
to a student.

Considering teachers only grade students towards the end of the module,
it is important to allow users to add modules to students without the
grade.

Let's
* add a ModuleCommand class to facilitate adding a module without a
  grade to the specified Person's list of modules.
* add a ModuleCommandParser class to parse the newly-added module
  command.
* refactor GradeCommand class such that it now
  * requires the specified Module to be present in the specified
    Person's list of modules.
  * sets the Grade field of the specified Module instead of adding a
    new Module object to the Person's list of modules.
* refactor GradeCommandParser class to support the changes made to the
  GradeCommand class.
* refactor Person class such that it supports the changes and additions
  made in this commit.

JavaDoc of GradeCommand and GradeCommandParser classes have also been
made as they were previously copies of the JavaDoc in the AddCommand and
AddCommandParser classes.