JsonAdaptedPerson: Add role handling
This commit updates JsonAdaptedPerson to handle roles in addition
to tags. A new List<JsonAdaptedRole> is added to store roles
ensuring that the roles field is serialized
to JSON and deserialized back into Role objects.

Changes include:
- Adding a roles field to JsonAdaptedPerson.
- Modifying constructors to handle roles.
- Updating toModelType() to convert JsonAdaptedRole back into Role
objects.
- Ensuring roles are included when constructing a Person object.

This change is necessary to ensure that the newly added Role field in th
e Person
model is correctly handled during JSON serialization and deserialization.