Provide temporary Person functionality
Current test coverage fails as Logic and Parser and etc have not
yet been updated to account for Patient and Doctor classes. The Id
class is giving errors when operating.

For the time being, the original Person class will be used in-place of
Patient and Doctor instances.

However, for future milestones, beyond
the MVP, let's:
*Make Person an abstract class
*Update all other parts of the code that relied on Person