Add id field to Event and Volunteer
Volunteer and Event now have an id field.

GUI displays their id after the name in the following format: (#ID)

Currently ID is not saved in a .json file, and every volunteer and
event will be given a new id in running number each time the
application is booted up. The feature to save IDs in a .json file
can be implemented in the future if deemed necessary.