Change implementation of opening browser from app
Previously java abstract window toolkit api was used to open a
browser from the javafx application. However, there were forum
posts stating that it might not be the best way to implement such
a function due to issues that may arise on mac OS. Thus the
implementation was changed to now use javafx's Application api.