Update the index to be group specific for Mark Task Command and Delete Task from Group
In the MarkTask and DeleteTaskFromGroup commands, they have an index
parameter that refers to the task to update. Previously,
the execute method for both commands made the index refer to the
position of the task in a list of all tasks globally instead of the
position in a list of tasks for a specific group.

Since being group specific is more intuitive as there is also a
parameter for a group name, the index now refers to the position of a
task in a list of tasks for a specific group.