Update task in Group's set of tasks for MarkTask Command
When calling MarkTaskCommand, it updated the status of the task in the
list of all tasks, but not in the group specific list of tasks. This
causes the status change to not be reflected when listing out the tasks
for a specific group.

Let's update the command to change the status for both the global and
group-specific list of tasks.